implement almost "unbeatable" opponent in the azure function
On 3x3 boards it is stronger than the DQN or Menace agents... wow. I just coded some moves that I find fitting for the logic of the game. Menace runs out of ideas/moves because it cannot win. DQN doesn't win often in the training phase, but at least it tries and wins randomly from to time. I wonder whether a randomly acting player would almost be as good as DQN or Menace. Or, maybe I did something wrong in the implementation, who knows.